+++ libraries/Utility/include/Tuple.hpp
@@ -9,7 +9,7 @@ namespace utility {
 
 class Tuple {
 public:
-    explicit Tuple(double x, double y, double z, double w): data{x,y,z,w}{};
+    explicit Tuple(float x, float y, float z, float w): data{x,y,z,w}{};
     Tuple() = default;
 
     bool isVector() const;
@@ -18,30 +18,30 @@ public:
     bool operator==(const Tuple& rhs)  const;
     Tuple operator+(const Tuple& rhs)  const;
     Tuple operator-(const Tuple& rhs)  const;
-    Tuple operator*(const double& rhs) const;
-    Tuple operator/(const double& rhs) const;
+    Tuple operator*(const float& rhs) const;
+    Tuple operator/(const float& rhs) const;
     Tuple operator-() const;
 
-    double magnitude() const;
+    float magnitude() const;
     Tuple normalize() const;
 
-    double dot(const Tuple& rhs) const;
+    float dot(const Tuple& rhs) const;
     Tuple cross(const Tuple& rhs) const;
     Tuple reflect(const Tuple& reflect) const;
     
-    const double& x() const;
-    double& x();
-    const double& y() const;
-    double& y();
-    const double& z() const;
-    double& z();
-    const double& w() const;
-    double& w();
+    const float& x() const;
+    float& x();
+    const float& y() const;
+    float& y();
+    const float& z() const;
+    float& z();
+    const float& w() const;
+    float& w();
 
-    std::array<double, 4> data;
+    std::array<float, 4> data;
 };
-Tuple Point(double x,double y, double z);
-Tuple Vector(double x,double y, double z);
+Tuple Point(float x,float y, float z);
+Tuple Vector(float x,float y, float z);
 
 std::ostream& operator<<(std::ostream& os, const Tuple& rhs);
